Allocate RocketListeners contiguously in bindings
Instead of using unique_ptrs and allocating RocketListener objects in
potentially different areas of memory, they should all be contiguously
allocated in the EdenRocketBindings vector of listeners.
To support this, new listeners are added via emplace_back instead of
push_back, and both a copy and move constructor were added to
RocketListener. By doing this, each construction of the RocketListener
would appropriately register a listener, so that even if a listener was
moved it would still be listening as expected.