Using 'emplace' instead of pushing temp objects
There are a lot of occasions where the EDEn code constructs a temporary
object and then immediately pushes it into a vector or map. Using
emplace to construct objects in-place in containers removes the need
for a temporary object.
Where it made sense, I replaced instances of “push_back”, “push” and
“insert” calls with their ‘emplace’ counterparts.