Storing Roster Characters as non-pointer members
CharacterRoster’s map of Characters now allocates them in-place in the
m_allCharacters map, instead of as raw pointers in the map.

Also fixed a bug where a message pipe reference is leaked on
destruction of the TileEngine, causing problems if leaving the
TileEngine and then going to a different state that updates the
PlayerData roster.