Extracting PlayerData from GameContext
Removing GameContext::getCurrentPlayerData() by using dependency
injection in the constructors of GameState classes that expect player
data. This is a little odd because I just did quite a bit of work to
move PlayerData to GameContext just to avoid the extra parameter, but
it’s for good reasons.
Prior to this change, there was one “current” PlayerData instance
constructed in the GameContext. From a design standpoint, this made
MainMenu kind of an odd state because it had a PlayerData, but had to
modify or reset the state of that data any time a TileEngine was either
created or destroyed. Since MainMenu doesn’t depend on a PlayerData
(and neither does SettingsMenu, for that matter), it doesn’t make sense
that the lifetime of PlayerData precedes them.
The other benefit of this approach is the ability to separate
SaveLocation from PlayerData. PlayerData didn’t really do anything with
SaveLocation, and it didn’t make sense to persist it in the PlayerData
class after its initial load. Instead, we will soon be able to create a
loader that gets both the SaveLocation and PlayerData from any given
save file, and uses the SaveLocation only on game load.