Saving game now saves player location
Removed “valid” flag from SaveLocation, replacing it with an isValid
function that checks for valid location data.
Removed SaveLocation instance from PlayerData. It’s now passed in on
PlayerData::save and returned inside the result of PlayerData::load.
PlayerData::load now returns a PlayerData instance and the SaveLocation
from the save game file.
Cleaned up how TileEngine initializes itself. Constructing a TileEngine
now *requires* a chapter string or a save location, and it is not
possible for other classes to initialize a TileEngine without these.
DataMenu has been augmented to take a SaveLocation, which can be easily
created and passed by the TileEngine.