Fixed failure to block NPC activation script
On activation, an NPC runs its activation function (if one is
available). If this function yields and blocks, then it should not
continue running right away. Fixed the NPC resume function to return
after calling its activation function, so that it doesn’t resume right
away if the activation function needs to wait on something (such as
dialogue or player action).