DataMenu becomes SaveMenu, a root-level menu
Since DataMenu is going to be used outside of the context of a regular
in-game menu, I’ve removed its inheritance of MenuState and made it a
top-level menu that doesn’t show a Shortcut bar or a Shell.
I decided to put all the DataMenu functionality I’m planning to create
into a subfolder of “Menu”. I’ll put the MainMenu directory in there
too in a future commit.