Removing workaround for CMake bug
Originally, CMake and Xcode didn’t play nice when creating a root
source group. In fact, it caused Xcode to crash when loading the
generated .xcodeproj. This no longer happens, either due to a CMake fix
or an Xcode fix. Either way, the check for Xcode before running this
script line is no longer needed.