Moving SaveMenu data logic to a Model object
SaveMenu was dealing with too much by being both a Model and a
Controller simultaneously. It required me to pass SaveMenu to anything
that needed to operate on save game data (such as the SaveMenu view
model). This has been fixed by splitting out the save game reading
logic and saving logic into a SaveGameModel class.