Supporting return values in async script functions
Plumbs through the ability for script functions (particularly
asynchronous ones) to return a set of results instead of always having
no return value.

This change introduces a CoroutineResults object which basically acts
as a tuple of return values for a given Task. When a Task is signaled
for completion, a component can choose to give it any amount of result
values, which it will then store. A Coroutine can then retrieve and
iterate over the results before resuming (for example, a Script can
push all the parameters onto the Lua stack as return values to a script
function call before resuming the Lua coroutine).

While doing this, also added waitForFinish support to the Actor Move
function. The function was used as a test bed for the new code, but it
has nothing meaningful to return. Still, the wait seems useful so I’m
keeping that in there. To facilitate this, I refactored some of the
functionality in ScriptFunctions.cpp so that it could be used by script
functions in other classes.