Holding Space no longer skips past dialogue
Previously, if you held Space to speed up the dialogue scrolling, it
would move on to the next line as soon as the end of the current line
was encountered. This is because Space is held and causing repeated
SDL_KEYDOWN events. Checking if the event is a repeat before moving on
to the next line fixes this issue.