Text event support, and updating debug console
Added text input support for Rocket elements using the SDL Text Input
event. This fixes the issue where the debug console didn’t respond to
text input at all.

Also updated the debug console to send a DebugCommandMessage through a
given message pipe instead of using the SDL event queue. This means the
event will be processed in the same frame, which is fine. The major
upshot of this is that the message is easier to create and interpret,
and doesn’t leave any dangling resources to clean up.