Separating dialogue box GUI from controller
Standard view/logic separation. Refactored all the view-specific parts
of DialogueController into a separate DialogueBox view, which is now
constructed outside of the controller and passed to it.

In the TileEngine’s case, the dialogue box is constructed in the tile
engine overlay, where the overlay itself now loads the RML and then
gives the dialogue box a specific RML element to manage.

Also tweaked the dialogue controller to create Line objects directly in
the line queue, instead of creating unique_ptrs, removing some of the
indirection there. To make sure that Lines were allocated and assigned
correctly, it helped to be more explicit about the rule that Tasks
can’t be copied, so I formalized this in the Task object.