Fixes TileEngine leak, crash on second game start
This was a fun one.

If you hit “New Game”, play a bit (or don’t), then press Escape to go
back to the Main Menu, then hitting “New Game” again causes a crash.
This happens because libRocket notices that the TileEngine Rocket
context wasn’t cleaned up, and so it doesn’t allocate a second one by
the same name. It returns null, causing a segmentation fault when the
context object is inevitably used.

The reason that the Rocket context is still hanging around is because
TileEngine doesn’t get cleaned up. Of course, since TileEngine is
allocated as a shared_ptr, the most likely culprit was a shared_ptr
cycle with TileEngine as a member.

However, upon searching the codebase, it seemed that all shared_ptr
instances that would hold the TileEngine are temporary parameters. Only
the ExecutionStack holds a shared_ptr to the TileEngine (in the state
queue), and the ScriptEngine only holds a weak_ptr to it.

Of course, after putting in a lot of breakpoints, and starting to
remove code, it became clear that the extra mystery reference was only
added if a script ran. And more specifically, the extra reference was
only added if a script coroutine was run and it called “setRegion”!

It turns out that setRegion called m_tileEngine.lock() to create a
local shared_ptr to TileEngine. Normally, this would clean up by
itself. Normally. TileEngine::setRegion can lead to a Map script being
called, which would ultimately cause the caller of setRegion to yield.
Lua’s yield function is implemented using a longjmp instruction, which
means that the program never returns to the end of setRegion’s call,
and thus never removes the shared_ptr instance!

So yes, boys and girls, even a temporary shared_ptr can leak. If you’re
going to call a function that has even a remote chance of calling
“longjmp”, make sure you’ve cleaned up your dangling resources and
shared_ptrs before making the call.

Changed the setRegion, narrate and say functions to get a raw pointer
to the TileEngine, allowing the temporary shared_ptr go out of scope,
and only *then* call the potentially yielding function inside the
TileEngine via the raw pointer.

Definitely one of the most subtle and interesting bugs I’ve seen, but
I’d love to have those hours of my life back.