Adding choice list support to dialogues
TL;DR — this change adds support for, and an example of, choice menus
in dialogue boxes, along with mouse support for choice selections.

This change adds “choices” as a parameter to the ‘say’ and ‘narrate’
functions in Lua. These functions can now be used to take a string
array of choices to present to the user. Many of the changes in this
commit are plumbing this “choices” parameter through to the
DialogueController, which constructs DialogueEntry objects with
optional choice lists, and ensuring that entries that present choices
can’t be skipped past without a choice first being selected by the
player.
This change also makes all the necessary UI changes to support choices
appearing in the dialogue box, as well as mouse input for selecting one
of the choices that appears. This includes adding the choice menu to
the RML, adding input handlers to the view class, making the dialogue
box operate on a div inside the overlay document instead of the entire
document itself (since it will support other UI pieces in the future),
as well as some additional RCSS polish. Finally, a ChoicesDataSource
was introduced so that the choices wouldn’t need to be added to the RML
manually by the DialogueBox class.