Use std::move to assign shared_ptrs passed by copy
A lot of method signatures in the code indicate that a shared_ptr needs
to be passed to them by copy. Although this constructs an additional
shared_ptr instance, it clarifies the intention of these functions to
hold ownership of these resources after the execution of the function
(e.g. constructors). However, it’s a little less than
optimal since the shared_ptr copying causes atomic
increments/decrements. A fix to this issue is to use std::move on the
shared_ptr copies when storing
them to reduce the cost of each shared_ptr field assignment.