Changing debugFlag vars to preprocessor macros
Every time a debugFlag was declared in a file with no DEBUG statement,
it’d trigger a compiler warning. At least 22 warnings would come out of
this alone.

To prevent the actual creation of a (potentially unused!) global
variable, changing the debugFlag declarations to #defines so that they
are simply inlined.