Separate Transition logic from TransitionState
Transitions (fades, blends, etc) are already able to take arbitrary
Textures to transition between. However, the next step is to use these
Transitions in things other than GameState transitions. This commit is
the first step in allowing that. The primary change is creating a
standalone “Transition” object, which the TransitionState defers to.
The Transition object is not a GameState subclass, but does support
“step” and “draw” functions for animation.