Fixing small logic error in DialogueEntry
As a safety precaution, DialogueEntry::removeNextScriptString was
changed to do nothing if there are no embedded scripts in the entry.
This was impossible anyway since the only caller checks against this
case before calling, but since that could easily change, it’s best for
this function to be more defensive.

DialogueEntry::getNextBracketPair was changed to expose only the
opening bracket index, and to return normally instead of by-ref.