TransitionState's Transition should be unique_ptr
Since TransitionState’s Transition instance isn’t meant to be shared,
it ought to be a unique_ptr anyway. While making this change and
updating clients of TransitionState, made a helper class to reduce the
amount of boilerplate code required to construct a TransitionState
(e.g. constructing a Transition with all its arguments and then moving
it into the TransitionState along with TransitionState’s arguments).