Reorganizing/optimizing header includes
Turns out that you *can* forward-declare types used in unique_ptrs as
long as the appropriate header is included before the destructor for
the unique_ptr owner is declared. In other words, pushing the
destructor down to the .cpp file allowed for forward declaration in
more places, which speeds up the build by a few seconds. ;)

Other changes included removing an unused function, and additionally
removing the unique_ptr usage for Spritesheet::m_texture.