Separate Party from all encountered characters in Roster data
Until now, we’ve assumed that all encountered characters are in the
party, and the party leader is always the first character in the party.
This commit modifies the savegame structure to make a Characters table
and a separate Party array that indexes into it via character ID. A
Leader key was also added to the savegame to properly persist the party
leader.