Adding collision shapes to Town Tileset
Taking advantage of Tiled's new Collision Editor to encode more collision data in the Tileset and remove some of the tedious work of setting up collisions for tiles that serve as obstacles.

Since more collision information is encoded in the tileset itself, many of the collision rectangles explicitly placed in the Tiled maps’ collision layers were no longer necessary, and were therefore removed from the maps.