Modify map Passibility drawing to blend with map instead of hiding it
Prior to this change, DRAW_PASSIBILITY would change a Map into a grid
of green/red tiles. Very easy to understand, but unfortunately
difficult to check for accuracy since the entire map, including the
obstacles impeding passibility, is obscured.

The new DRAW_IMPASSIBILITY branch draws the map regularly, but adds a
red highlight over top of tiles that are impassible. This change makes
it a lot easier to ensure that collision boxes are consistent with
actual obstacles on the map.