Refine GameState frame delay
The previous solution of adding a delay before advancing the frame
didn’t actually alter `timePassed` to account for the additional time
passed due to the SDL_Delay call. The difference was smalll (<16ms) but
still something that could build up and visibly mess up the amount of
time the game thinks has passed very quickly.
This commit updates `timePassed` to account for the SDL_Delay.