Switching to GLEW instead of hand-rolling extension-wrangling
As EDEn begins to transition to more modern OpenGL, more extensions
will be used and the `OpenGLExtensions` object is going to get pretty
hairy to continue maintaining. GLEW already does a lot of the work that
would be reinvented in `OpenGLExtensions`, so it’s better to just use
it instead.