Rename Actor* classes to GridActor*
Sometimes the most annoying part of a refactor is the renaming, because
so many files are affected!
Many parts of GridActor will be pulled up to an Actor superclass that
can be used by the battle system to position and control combatant
sprites, but this commit stops at the rename to keep commits as small
as possible.