Fix scheduler switch issue if script function changes GameState
If a script function’s work causes a GameState switch (pushing a new
state like a battle or a menu, or popping an existing state), then this
changes the scheduler being used.
This means that the old script engine code potentially started a Task
on one scheduler, and then blocked the running coroutine on another
scheduler (which doesn’t make sense).
This leads to undesirable behavior, such as potential segfaults if the
new GameState has just been constructed and doesn’t actually have a
running coroutine yet. If there is a running coroutine, it’s the wrong
one to block, which will likely lead to deadlock.

The fix here is to change waitUntilFinished into a function that
creates the task, executes the work AND blocks on it, performing all
those steps using the same Scheduler instance.