slideCamera should use a Task to signal completion
TileEngine::slideCamera used Scheduler::join internally to always block
the running coroutine. This meant that the script engine had no
opportunity to determine how the slide should affect execution of the
running corotuine. As well, it returned 0 to signal to Lua functions
that it’s a void function, but it shouldn’t have to worry about Lua
implementation details at all.
Like the Save Menu, the Camera Slider has been altered to use and
signal a task instead, and the slideCamera script API function itself
now handles waiting on the slide.

Finally, removed the origin parameter from the slider since it wasn’t
being used and was unlikely to be used since it would cause the camera
to instantly jump to a location before starting a smooth slide.