Add Lua-based Character init script
Character instances are now initialized after construction by being
passed to a Lua-based character metadata script that allows rules about
character initialization to be customized outside of the engine itself.

The running of the script is done by Metadata, but since running a
script is not const, it means that Metadata refs in PlayerData objects
need to be non-const now.