Allow for placing NPCs using Tiled Map Editor
While it’s extremely helpful to expose NPCcreation to Lua scripts, it’s
a pain to have to create NPCs without being able to directly visualize
where on the map they are. This change adds support for an object layer
called ‘npcs’.
In addition, added a string->Direction conversion helper. It would be
nice to be able to genericize this to all enums, but that’s likely
non-trivial.