Fail gracefully when quest insertion fails
Encountered a segfault in the LuaQuest addQuest script hook. If there is an attempt to create a quest that already exists, then the insertion fails, which causes the duplicate Quest to be returned while simultaneously being cleaned up (since it isn't in the Quest log, there are no remaining references and the shared_ptr resource is cleaned up).

Fixed this by passing the result of the insertion to the script glue code, so it can pass an error to the script.