Don't persist fast dialogue mode after a line
When Space is pressed and then released at the end of a dialogue line (to advance the dialogue), fast mode can turn on and then fail to turn off as the dialogue box gets emptied out and loses keyboard focus. To fix this, this commit causes fast mode to proactively deactivate on every key up event, even if there is no dialogue queued up.